returned amisto's script and summed up all my contributions (shells)
