+++ tools/ndt/geomaker_curve_shells.cpp
